FIX montage reorganization
